+++ ionoscloud/test_constants.go
@@ -1664,9 +1664,13 @@ resource "random_string" "simple_string" {
 }
 `
 
+// K8s values
+
 const K8sVersion = "1.26.4"
 const UpgradedK8sVersion = "1.26.6"
-const K8sBucket = "sdktestv6"
+const K8sBucket = "test_k8s_terraform_v7"
+const K8sPrivateClusterNATGatewayIP = "85.215.52.107"
+const K8sPrivateClusterNodeSubnet = "192.168.0.0/16"
 
 // DNS test constants: configs, attributes and values.